anki-voice-server/pyproject.toml
[tool.poetry]
name = "anki-voice-server"
version = "0.1.0"
description = "FastAPI server for Anki Voice system"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
httpx = "^0.25.2"
python-dotenv = "^1.0.0"
beautifulsoup4 = "^4.12.2"
pydantic = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

